Refactored DocumentCategory Karma test to Jest test / Removed karma test
